requirements.txt
flask
python-dotenv
psycopg2-binary
flask-cors
flask-jwt-extended